+++ app/data_migrations/populate_taxonomy.py
@@ -1,100 +1,8 @@
-import json
-from typing import Sequence
-
 from sqlalchemy.orm import Session
+from app.data_migrations.taxonomy_cclw import get_cclw_taxonomy
 
 from app.db.models.app.users import Organisation
 from app.db.models.law_policy.metadata import MetadataOrganisation, MetadataTaxonomy
-
-"""At the moment taxonomy is kept simple, and only supports string validation for enums
-
-For Example:
-
-{
-    "topic": {
-       allowed_values: [],
-       allow_blanks: false,
-    },
-    ...
-}
-
-"""
-
-TAXONOMY_DATA = [
-    {
-        "key": "topic",
-        "filename": "app/data_migrations/data/cclw/topic_data.json",
-        "file_key_path": "name",
-        "allow_blanks": True,
-    },
-    {
-        "key": "sector",
-        "filename": "app/data_migrations/data/cclw/sector_data.json",
-        "file_key_path": "node.name",
-        "allow_blanks": True,
-    },
-    {
-        "key": "keyword",
-        "filename": "app/data_migrations/data/cclw/keyword_data.json",
-        "file_key_path": "name",
-        "allow_blanks": True,
-    },
-    {
-        "key": "instrument",
-        "filename": "app/data_migrations/data/cclw/instrument_data.json",
-        "file_key_path": "node.name",
-        "allow_blanks": True,
-    },
-    {
-        "key": "hazard",
-        "filename": "app/data_migrations/data/cclw/hazard_data.json",
-        "file_key_path": "name",
-        "allow_blanks": True,
-    },
-    {
-        "key": "framework",
-        "filename": "app/data_migrations/data/cclw/framework_data.json",
-        "file_key_path": "name",
-        "allow_blanks": True,
-    },
-]
-
-
-def dot_dref(obj: dict, dotted_key: str):
-    if "." not in dotted_key:
-        return obj[dotted_key]
-    keys = dotted_key.split(".", 1)
-    return dot_dref(obj[keys[0]], keys[1])
-
-
-def load_metadata_type(filename: str, key_path: str) -> Sequence[str]:
-    with open(filename) as file:
-        data = json.load(file)
-    return [dot_dref(obj, key_path) for obj in data]
-
-
-def get_cclw_taxonomy():
-    taxonomy = {}
-    for data in TAXONOMY_DATA:
-        taxonomy.update(
-            {
-                data["key"]: {
-                    "allowed_values": load_metadata_type(
-                        data["filename"], data["file_key_path"]
-                    ),
-                    "allow_blanks": data["allow_blanks"],
-                },
-            }
-        )
-
-    # Remove unwanted values for new taxonomy
-    if (
-        "sector" in taxonomy
-        and "Transportation" in taxonomy["sector"]["allowed_values"]
-    ):
-        taxonomy["sector"]["allowed_values"].remove("Transportation")
-
-    return taxonomy
 
 
 def populate_org_taxonomy(
