+++ app/data_migrations/taxonomy_utils.py
@@ -1,0 +1,29 @@
+import json
+from typing import Sequence
+
+"""At the moment taxonomy is kept simple, and only supports string validation for enums
+
+For Example:
+
+{
+    "topic": {
+       allowed_values: [],
+       allow_blanks: false,
+    },
+    ...
+}
+
+"""
+
+
+def dot_dref(obj: dict, dotted_key: str):
+    if "." not in dotted_key:
+        return obj[dotted_key]
+    keys = dotted_key.split(".", 1)
+    return dot_dref(obj[keys[0]], keys[1])
+
+
+def load_metadata_type(filename: str, key_path: str) -> Sequence[str]:
+    with open(filename) as file:
+        data = json.load(file)
+    return [dot_dref(obj, key_path) for obj in data]
